calculate time

+++ src/test/Sorter.java
@@ -3,7 +3,7 @@ package test;
 import simplesorting.Swaping;
 
 public class Sorter {
-    public static void bubbleSort(int[] array){
+    public static void bubbleSort(long[] array){
         for(int out=array.length-1; out>0; out--){
             for(int in=0; in<out; in++){
                 if(array[in]>array[in+1])
@@ -12,7 +12,7 @@ public class Sorter {
         }
     }
 
-    public static void selectionSort(int[] array){
+    public static void selectionSort(long[] array){
         for(int out=0; out<array.length; out++){
             int min = out;
             for(int in=out+1; in<array.length; in++)
@@ -22,9 +22,9 @@ public class Sorter {
         }
     }
 
-    public static void insertionSort(int[] array){
+    public static void insertionSort(long[] array){
         for(int out=1; out<array.length; out++){
-            int temp = array[out];
+            long temp = array[out];
             int in;
             for(in=out; in>0 && array[in-1]>temp; in--){
                 array[in] = array[in-1];
